Feat/user store detail

+++ home/domain/src/main/java/com/home/domain/data/store/FeedbackModel.kt
@@ -3,5 +3,5 @@ package com.home.domain.data.store
 data class FeedbackModel(
     val count: Int = 0,
     val feedbackType: FeedbackType,
-    val ratio: Int = 0
+    val ratio: Double = 0.0
 )
